devel/gettext-tools: allow to specify an archive format
devel/gettext-tools uses xz to create an archive that is added to its
package. A user can pass a flag to CONFIGURE_ARGS to specify a
non-default archive format, which in result breaks a package build that
assumes the .tar.xz format in the pkg-plist file.

Allow to specify the format with ARCHIVE_SUFX that can be used by the
package building infrastructure to use bzip2 instead of xz to create the
archive.